Should cast IDs to unsigned longs before printing on x64.

+++ main.c
@@ -93,7 +93,7 @@ void ssHandleConnectTCP(socketServer *server, socketDetails *details){
 	          (void *)(&((struct sockaddr_in6 *)&details->address)->sin6_addr)),
 	          &IP[0],
 	          sizeof(IP));
-	printf("Accepted TCP connection from %s:%u (socket #%u).\n", IP, ((struct sockaddr_in *)&details->address)->sin_port, details->id);
+	printf("Accepted TCP connection from %s:%u (socket #%lu).\n", IP, ((struct sockaddr_in *)&details->address)->sin_port, (unsigned long)details->id);
 }
 
 void ssHandleBufferTCP(const socketServer *server, socketDetails *details){
@@ -104,7 +104,7 @@ void ssHandleBufferTCP(const socketServer *server, socketDetails *details){
 	          (void *)(&((struct sockaddr_in6 *)&details->address)->sin6_addr)),
 	          &IP[0],
 	          sizeof(IP));
-	printf("Data received over TCP from %s:%u (socket #%u): %s\n", IP, ((struct sockaddr_in *)&details->address)->sin_port, details->id, details->lastBuffer);
+	printf("Data received over TCP from %s:%u (socket #%lu): %s\n", IP, ((struct sockaddr_in *)&details->address)->sin_port, (unsigned long)details->id, details->lastBuffer);
 	ssSendDataTCP(details->handle, "Data received over TCP successfully. You should get this.\n");
 }
 
@@ -116,7 +116,7 @@ void ssHandleDisconnectTCP(socketServer *server, socketDetails *details){
 	          (void *)(&((struct sockaddr_in6 *)&details->address)->sin6_addr)),
 	          &IP[0],
 	          sizeof(IP));
-	printf("Closing TCP connection with %s:%u (socket #%u).\n", IP, ((struct sockaddr_in *)&details->address)->sin_port, details->id);
+	printf("Closing TCP connection with %s:%u (socket #%lu).\n", IP, ((struct sockaddr_in *)&details->address)->sin_port, (unsigned long)details->id);
 	scRemoveSocket(&server->connectionHandler, details);
 }
 
@@ -128,7 +128,7 @@ void ssHandleConnectUDP(socketServer *server, socketDetails *details){
 	          (void *)(&((struct sockaddr_in6 *)&details->address)->sin6_addr)),
 	          &IP[0],
 	          sizeof(IP));
-	printf("Accepted UDP connection from %s:%u (socket #%u).\n", IP, ((struct sockaddr_in *)&details->address)->sin_port, details->id);
+	printf("Accepted UDP connection from %s:%u (socket #%lu).\n", IP, ((struct sockaddr_in *)&details->address)->sin_port, (unsigned long)details->id);
 }
 
 void ssHandleBufferUDP(const socketServer *server, socketDetails *details){
@@ -139,7 +139,7 @@ void ssHandleBufferUDP(const socketServer *server, socketDetails *details){
 	          (void *)(&((struct sockaddr_in6 *)&details->address)->sin6_addr)),
 	          &IP[0],
 	          sizeof(IP));
-	printf("Data received over UDP from %s:%u (socket #%u): %s\n", IP, ((struct sockaddr_in *)&details->address)->sin_port, details->id, details->lastBuffer);
+	printf("Data received over UDP from %s:%u (socket #%lu): %s\n", IP, ((struct sockaddr_in *)&details->address)->sin_port, (unsigned long)details->id, details->lastBuffer);
 	ssSendDataUDP(&server->connectionHandler, details, "Data received over UDP successfully. You might get this.\n");
 }
 
@@ -151,7 +151,7 @@ void ssHandleDisconnectUDP(socketServer *server, socketDetails *details){
 	          (void *)(&((struct sockaddr_in6 *)&details->address)->sin6_addr)),
 	          &IP[0],
 	          sizeof(IP));
-	printf("Closing UDP connection with %s:%u (socket #%u).\n", IP, ((struct sockaddr_in *)&details->address)->sin_port, details->id);
+	printf("Closing UDP connection with %s:%u (socket #%lu).\n", IP, ((struct sockaddr_in *)&details->address)->sin_port, (unsigned long)details->id);
 	scRemoveSocket(&server->connectionHandler, details);
 }
